% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplepx-package.R
\docType{package}
\name{simplepx}
\alias{simplepx}
\title{SIMPLEPX}
\description{
An R package for accessing PX-Web APIs. The package might work with other
PX-Web APIs as well, but so far has only been tested with the API of
Statistics Finland.
}
