% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px.R
\name{doc-all}
\alias{doc-all}
\alias{px_nav}
\alias{px_var}
\alias{px_dl}
\title{Download data from the PX-Web API of Statistics Finland}
\usage{
px_nav(path = "", api = "https://pxnet2.stat.fi/PXWeb/api/v1/en/")

px_var(path, api = "https://pxnet2.stat.fi/PXWeb/api/v1/en/")

px_dl(path, var = NULL, simplify_strings = FALSE, na_omit = FALSE,
  api = "https://pxnet2.stat.fi/PXWeb/api/v1/en/")
}
\arguments{
\item{path}{suffix to the path after \code{api}.
For \code{px_download} and \code{px_var} this must point to an actual
dataset (ie. an url ending in .px), whereas for \code{px_nav} this should
correspond to a partial path to a dataset (ie. an url that points to a
directory). Defaults to \code{""}.}

\item{api}{The url to the API. Defaults to
\code{"https://pxnet2.stat.fi/PXWeb/api/v1/en/"}, the PX-Web API of
Statistics Finland.}

\item{var}{The values of the dimensions to be queried as a tibble. Can be
obtained by filtering from the result of \code{px_var}. If \code{NULL}, the
defaults are selected.}

\item{simplify_strings}{If \code{TRUE}, the strings in the result (colnames,
values of the dimensions) are set to all lowercase, all special characters
are removed and spaces are replaced by underscores. Defaults to
\code{FALSE}.}

\item{na_omit}{If \code{TRUE}, all missing values are omitted.
Defaults to \code{FALSE}.}
}
\value{
\code{px_dl} and \code{px_var} always return a tibble.
 \code{px_nav} returns a tibble when request is sent to a non-table URL
 (ie. an url corresponds to a directory rather than a particular data set).
}
\description{
\code{px_nav} is used to navigate the API: it returns the possible
 subdirectories as a tibble with dbid corresponding to the suffix in the url
 that points to the given subdirectory.
}
\details{
\code{px_var} returns the values of the dimensions in the given table and
 \code{px_dl} returns the actual data corresponding to variables specified
 as the \code{var} argument. The format of the \code{var} argument is the
 same as in the tibble returned by \code{px_var}, so the values can be
 selected by filtering that tibble. See the examples for mode information.
}
\examples{
\donttest{
# Navigate through the API
px_nav()
# check what data is available under StatFin
px_nav("StatFin/")
# navigate through the directories,
# check what variables are available for population statistics
var <- px_var("StatFin/vrm/synt/statfin_synt_pxt_011.px")
# select years after 1900
var_1900 <- dplyr::filter(var, Year >= 1900)
# Download the data, starting from year 1900,
# omitting the var-argument would download the data for all the years.
data <- px_dl("StatFin/vrm/synt/statfin_synt_pxt_011.px", var_1900,
              simplify_strings = TRUE)
}

}
